Update README.md
People use emojis each day. Emojis have become a new language that can more effectively express an idea or emotion. 
This visual language is now a standard for online communication, available not only in Twitter, but also in another large online platform such as Facebook and Instagram. 

In Today’s Generation people usually tend to communicate with each other using Emoticons. So, we thought of making our own customized emojis.
Emojify is a software which deals with the creation of Emoji’s or Avatars. The neural network has been an emerging application in numerous and diverse areas as example of end to end learning .
This paper is based on a system which implements Convolutional Neural Network and Fer Dataset to detect emotions from facial expressions and converting them to personalized emojis.